quarantine_animals_service/pyproject.toml
[project]
name = "quarantine-animals-service"
version = "0.1.0"
description = "Module: Quarantine Animals - FastAPI backend with Postgres"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.9.0",
  "psycopg2-binary>=2.9.9",
  "python-dotenv>=1.0.1",
  "typing-extensions>=4.12.2",
  "python-multipart>=0.0.9",
  "httpx>=0.27.0",
]

[tool.pytest.ini_options]
addopts = "-q"
